BAC-5-3: add refactor modal detail
