+++ src/features/Module/PrepareToQuize/ui/ModuleDetail/constants.ts
@@ -1,0 +1,4 @@
+export const MODAL = {
+  ID: 'detail-module',
+  TITLE: 'Module detail'
+};
