+++ src/features/Module/PrepareToQuize/ui/ModuleDetail/index.ts
@@ -1,0 +1,2 @@
+export { default as ModuleDetail } from './ModuleDetail.vue';
+export * from './constants';
